feat: add safe mode option to AsciidoctorOptions
This commit introduces a new property 'safe' to the AsciidoctorOptions
interface. The 'safe' property can take one of the following values:
'safe', 'unsafe', 'server', or 'secure'. This option is then passed to
the Asciidoctor.convert method to control the safe mode for rendering.

The package version is also bumped to 0.2023.7.