feat: Add Asciidoctor WASM conversion in index.html
This commit introduces a new feature in the index.html file. It adds
Asciidoctor WASM conversion functionality. The user can input Asciidoc
content in a textarea, and upon clicking a button, the content is
converted to HTML and displayed in an iframe. This is achieved by
importing and initializing the Asciidoctor WASM module in a script tag.
The commit also includes responsive styling for the layout.