feat(asciidoctor): support true value in attributes and refactor options
This commit introduces the ability to use a boolean true value in the
attributes of AsciidoctorOptions. It also refactors the way options are
handled in the convert method of Asciidoctor class. The changes include
a more explicit handling of different types of values in the options.

Tests have been updated to reflect these changes, including a new test
for the use of attributes with a true value.