feat: Update dependencies and refactor RubyVM setup
This commit updates the dependencies in deno.json and refactors the
RubyVM setup in the project. The import path for AsciidoctorWASM in
benchmark.ts has been updated. The version of the project in deno.json
has been bumped up. The import paths for DefaultRubyVM in browser.ts
and node.ts have been updated to point to a new file ruby.ts. A new
dependency uwasi has been added in deno.json. A new file ruby.ts has
been created which sets up the RubyVM using uwasi instead of the
previous setup.